device-gui/requirements.txt
pynacl==1.5.0
requests==2.31.0
